+++ src/H2DE_abs_rect.cpp
@@ -1,0 +1,5 @@
+#include "H2DE_utils.h"
+
+H2DE_AbsRect::operator SDL_Rect() const {
+    return { x, y, w, h };
+}