+++ src/H2DE_abs_size.cpp
@@ -1,0 +1,6 @@
+#include "H2DE_utils.h"
+
+// METHODS
+H2DE_AbsRect H2DE_AbsSize::makeRect(H2DE_AbsPos pos) const {
+    return { pos.x, pos.y, w, h };
+}